🧬 Migrate the data model *IteratorItem* to module *api_config.item*.
